feat: refactor contact management components and services, replace Button and Container components and implement SaveBtn and AddBtn
