+++ apps/mobile/app/+not-found.tsx
@@ -1,7 +1,7 @@
 import { Link, Stack } from 'expo-router';
 import { Text } from 'react-native';
 
-import { Container } from '~/components/Container';
+import Container from '~/components/shared/Container';
 
 export default function NotFoundScreen() {
   return (
