+++ apps/mobile/app/_layout.tsx
@@ -6,7 +6,7 @@ export default function Layout() {
   return (
     <Stack>
       <Stack.Screen name="index" options={{ title: 'Lazor Connect' }} />
-      <Stack.Screen name="details" options={{ title: 'Details' }} />
+      <Stack.Screen name="contact/create" options={{ title: 'Create contact' }} />
     </Stack>
   );
 }
