--- apps/mobile/app/details.tsx
@@ -1,13 +1,0 @@
-import { Text } from 'react-native';
-
-import { Container } from '~/components/Container';
-
-export default function Details() {
-  return (
-    <>
-      <Container>
-        <Text>Hola</Text>
-      </Container>
-    </>
-  );
-}
