+++ apps/mobile/app/index.tsx
@@ -1,14 +1,10 @@
-import { Link } from 'expo-router';
-
-import { Button } from '~/components/Button';
-import { Container } from '~/components/Container';
+import AddBtn from '~/components/contacts/AddBtn';
+import Container from '~/components/shared/Container';
 
 export default function Home() {
   return (
     <Container>
-      <Link href="/contact/create" asChild>
-        <Button title="Añadir contacto" />
-      </Link>
+      <AddBtn />
     </Container>
   );
 }
