+++ apps/mobile/components/contacts/AddBtn.tsx
@@ -1,0 +1,14 @@
+import { Link } from 'expo-router';
+import { Plus } from 'lucide-react-native';
+
+import Button from '~/components/shared/Button';
+
+export default function AddBtn() {
+  return (
+    <Link href="/contact/create" asChild>
+      <Button className="absolute bottom-0 right-0 h-20 w-20 rounded-3xl">
+        <Plus size={32} color="white" />
+      </Button>
+    </Link>
+  );
+}
