+++ apps/mobile/components/shared/Container.tsx
@@ -1,5 +1,5 @@
 import { SafeAreaView } from 'react-native';
 
-export const Container = ({ children }: { children: React.ReactNode }) => {
+export default function Container({ children }: { children: React.ReactNode }) {
   return <SafeAreaView className="m-6 flex flex-1">{children}</SafeAreaView>;
-};
+}
