+++ apps/mobile/package.json
@@ -22,6 +22,7 @@
     "expo-status-bar": "~2.2.3",
     "expo-system-ui": "~5.0.6",
     "expo-web-browser": "~14.1.6",
+    "lucide-react-native": "^0.511.0",
     "nativewind": "latest",
     "react": "19.0.0",
     "react-dom": "19.0.0",
